Collector interface with Query struct
Established the Collector interface with a corresponding query struct.
Added the parameterType enum in order to keep track of what
parameters types of parameters the query object will support.

Currently, the the process of adding a new parameter would be to
add a corresponding field to the query struct and append onto the
parameterType enum. The following commit will provide an example
of this.